requirements.txt
Django==4.0.3
django-livereload-server==0.4
gunicorn==20.1.0
django-import-export==2.8.0
Pillow==9.0.1
djangorestframework==3.13.1
django-environ==0.8.1
django-grappelli==3.0.3